make it possible to check recordings
with 

pushover -c recordFile

it is now possible to check recordings. The level must have ended
right on that spot. This is now happening without any use of graphics
and quite quick

this made it necessary to initialize the ant arrays with proper sizes
otherwise the animations would not be properly simulated